[planning-cpp] Update test settings.

+++ fire_rs/planning-cpp/src/utils.h
@@ -30,13 +30,13 @@ double positive_modulo(double left, double right) {
 }
 
 /** Macro that test equality of to floating point number, disregarding rounding errors. */
-#define ALMOST_EQUAL(x, y) (fabs(x - y) < 0.000001)
+#define ALMOST_EQUAL(x, y) (fabs((double) x - (double) y) < 0.000001)
 
 /** Macro that test whether x >= y with tolerance to rounding errors. */
-#define ALMOST_GREATER_EQUAL(x, y) (x >= (y - 0.000001))
+#define ALMOST_GREATER_EQUAL(x, y) ((double) x >= ((double) y - 0.000001))
 
 /** Macro that test whether x >= y with tolerance to rounding errors. */
-#define ALMOST_LESSER_EQUAL(x, y) (x <= (y + 0.000001))
+#define ALMOST_LESSER_EQUAL(x, y) ((double) x <= ((double) y + 0.000001))
 
 #define ASSERT(test)                  \
 if(!(test)) {                         \
